+++ src/core/chart-options/DonutChartConfiguration.class.ts
@@ -1,11 +1,14 @@
 import { PieChartConfiguration } from './PieChartConfiguration.class';
 
 export class DonutChartConfiguration extends PieChartConfiguration {
+
   readonly type: string = 'donutChart';
   readonly donut: boolean = true;
-  donutRatio: number = 0.35;
+
+  public donutRatio: number = 0.35;
 
   constructor() {
     super();
   }
+
 }